CollectionType-class {GSEABase}	R Documentation

Class "CollectionType"

Description

These classes provides a way to tag the origin of a GeneSet. Collection types can be used in manipulating (e.g., selecting) sets, and can contain information specific to particular sets (e.g., category and subcategory classifications of BroadCollection.)

Objects from the Class

The following classes can tag gene sets; GO, KEGG, Chr, Chrloc, OMIM, and PMID collections can be derived from chip or organism ‘annotation’ packages.

NullCollection

No formal collection information available.

BroadCollection

Derived from, or destined to be, Broad XML. Usually created and written getBroadSets, toBroadXML.

ComputedCollection

A computationally created collection, e.g., by performing logic operations on gene sets.

ExpressionSetCollection

Derived from ExpressionSet. Usually created during a call to GeneSet or GeneColorSet.

GOCollection

Collection derived using Gene Ontology (GO) terms.

OBOCollection

Collection derived from GOCollection, specifically from files described by the OBO file format. See OBOCollection

KEGGCollection

Collection derived using KEGG terms.

ChrCollection

Collection derived using chromsome locations

ChrlocCollection

Collection derived using chromosome starting posistions

MapCollection

Collection derived from cytogenic bands.

OMIMCollection

Collection derived from identifiers in the Online Inheritance in Man.

PMIDCollection

Collection derived from PMID identifiers.

PfamCollection

Collection derived from Pfam identifiers.

PrositeCollection

Collection derived from Prosite identifiers.

Objects are instantiated with calls to CollectionType constructors, with slot names as possible arguments.

Slots

CollectionType classes (Null, ComputedCollection, ExpressionSet) have the slot:

type:

Object of class "ScalarCharacter" containing the character string representation of this CollectionType.

CollectionIdType classes (KEGG, OMIM, PMID, Chr, Chrloc, Map, GO) extend the CollectionType and have the additional slot:

ids:

Object of class "character" containing a vector of character string representations of corresponding identifiers, e.g., ‘KEGG’ or ‘GO’ terms.

GOCollection extends CollectionIdType and has the additional slot:

evidenceCode:

Object of class "character", containing GO evidence codes used to construct the gene set.

ontology

Object of class "character" vector of GO ontology terms used to filter GO terms in the GO Collection.

The values of evidenceCode are

Experimental Evidence Codes

EXP

Inferred from Experiment

IDA

Inferred from Direct Assay

IPI

Inferred from Physical Interaction

IMP

Inferred from Mutant Phenotype

IGI

Inferred from Genetic Interaction

IEP

Inferred from Expression Pattern

Computational Analysis Evidence Codes

ISS

Inferred from Sequence or Structural Similarity

ISO

Inferred from Sequence Orthology

ISA

Inferred from Sequence Alignment

ISM

Inferred from Sequence Model

IGC

Inferred from Genomic Context

RCA

inferred from Reviewed Computational Analysis

Author Statement Evidence Codes

TAS

Traceable Author Statement

NAS

Non-traceable Author Statement

Curator Statement Evidence Codes

IC

Inferred by Curator

ND

No biological Data available

Automatically-assigned Evidence Codes

IEA

Inferred from Electronic Annotation

OBOCollection extends GOCollection; see OBOCollection.

BroadCollection has slots:

category:

Object of class "ScalarCharacter" containing terms from the Broad list of categories, or NA

subCategory:

Object of class "ScalarCharacter" containing Broad sub-categories, or NA

Methods

CollectionType classes have methods:

collectionType<-

signature(object = "GeneSet", value = "CollectionType"): Replace the CollectionType

collectionType

signature(object = "CollectionType"): Retrieve the collection type.

|, &, intersect, union, setdiff

signature(e1="CollectionType", e2="CollectionType"): return e1 when class(e1) and class(e2) are the same, or ComputedCollection when different.

show

signature(object = "CollectionType"): display the collection type.

CollectionIdType classes inherit CollectionType methods, and have in addition:

ids

signature(object="CollectionIdType"): Retrieve the identifiers of the collection type.

[

signature(object="CollectionIdType", i="missing", j="missing", ..., ids=ids(object)): return a subset of object containing only ids in ids

|, &, intersect, union, setdiff

signature(e1="CollectionIdType", e2="CollectionIdType"): always return ComputedCollection.

GOCollection inherits CollectionIdType methods, and has in addition:

evidenceCode

Retrieve the evidence codes of the GO collection.

ontology

Retrieve the ontology terms of the GO collection.

[

signature(object="CollectionIdType", i="missing", j="missing", ..., evidenceCode=evidenceCode(object), ontology=ontology(object)): return a subset of object containing only evidence and ontology codes in evidenceCode, ontology. This method passes arguments ... to [,CollectionIdType methods.

BroadCollection has methods:

bcCategory

Retrieve the category of the Broad collection.

bcSubCategory

Retrieve the sub-category of the Broad collection.

Author(s)

Martin Morgan <Martin.Morgan@RoswellPark.org>

See Also

CollectionType consturctors; getBroadSets for importing collections from the Broad (and sources).

Examples


names(getClass("CollectionType")@subclasses)

## Create a CollectionType and ask for its type
collectionType(ExpressionSetCollection())

## Read two GeneSets from a Broad XML file into a list, verify that
## they are both BroadCollection's. Category / subcategory information
## is unique to Broad collections.
fl <- system.file("extdata", "Broad.xml", package="GSEABase")
sets <- getBroadSets(fl)
sapply(sets, collectionType)

## ExpressionSets are tagged with ExpressionSetCollection; there is no
## 'category' information.
data(sample.ExpressionSet)
gs <- GeneSet(sample.ExpressionSet[100:109],
              setName="sample.GeneSet", setIdentifier="123")
collectionType(gs)

## GOCollections are created by reference to GO terms and evidenceCodes
GOCollection("GO:0005488")
## requires library(GO); EntrezIdentifers automatically created
## Not run: 
GeneSet(GOCollection(c("GO:0005488", "GO:0019825"),
                     evidenceCode="IDA"))

## End(Not run)

[Package GSEABase version 1.44.0 Index]