GeneColorSet {GSEABase}	R Documentation

Methods to Construct "GeneColorSet" Instances

Description

GeneColorSet is a generic for constructing gene color sets (i.e., gene sets with "coloring" to indicate how features of genes and phenotypes are associated).

Methods

Available methods are the same as those for GeneSet, but a GeneColorSet requires an additional phenotype argument to identify the phenotype that is being colored. See documentation for GeneColorSet for examples.

An additional method is:

signature(type = "GeneSet", phenotype="character")

This method constructs a 'color' gene set from an uncolored gene set.

See Also

GeneColorSet-class

[Package GSEABase version 1.44.0 Index]