getOBOCollection {GSEABase}	R Documentation

Read OBO-specified Gene Ontology Collections

Description

getOBOCollection parses a uri (file or internet location) encoded following the OBO specification defined by the Gene Onotology consortium.

Usage


getOBOCollection(uri, evidenceCode="ANY", ...)

Arguments

uri
A file name or URL containing gene sets encoded following the OBO specification.

evidenceCode
A character vector of evidence codes.

...
Further arguments passed to the OBOCollection constructor.

Value

getOBOCollection returns an OBOCollection of gene sets. The gene set is constructed by parsing the file for id tags in TERM stanzas. The parser does not currently support all features of OBO, e.g., the ability to import additional files.

Author(s)

Martin Morgan <mtmrogan@fhcrc.org>

References

http://www.geneontology.org

See Also

OBOCollection, OBOCollection

Examples


## 'fl' could also be a URI
fl <- system.file("extdata", "goslim_plant.obo", package="GSEABase")
getOBOCollection(fl) # GeneSetCollection of 2 sets

## Not run: 
## Download from the internet
fl <- "http://www.geneontology.org/GO_slims/goslim_plant.obo"
getOBOCollection(fl, evidenceCode="TAS")

## End(Not run)

[Package GSEABase version 1.44.0 Index]