goSlim-methods {GSEABase}	R Documentation

Methods for Function goSlim in Package ‘GSEABase’

Description

These methods summarize the gene ontology terms implied by the idSrc argument into the GO terms implied by the slimCollection argument. The summary takes identifiers in idSrc and determines all GO terms that apply to the identifiers. This full list of GO terms are then classified for membership in each term in the slimCollection.

The resulting object is a data frame containing the terms of slimCollection as row labels, counts and frequencies of identifiers classified to each term, and an abbreviated term description.

An identifier in idSrc can expand to several GO terms, and the GO terms in slimCollection can imply an overlapping hierarchy of terms. Thus the resulting summary can easily contain more counts than there are identifiers in idSrc.

Usage


goSlim(idSrc, slimCollection, ontology, ..., verbose=FALSE)

Arguments

idSrc
An argument determining the source of GO terms to be mapped to slim terms. The source might be a GOCollection of terms, or another object (e.g., ExpressionSet) for which the method can extract GO terms.

slimCollection
An argument containing the GO slim terms.

ontology
A character string naming the ontology to be consulted when identifying slim term hierarchies. One of ‘MF’ (molecular function), ‘BP’ (biological process), ‘CC’ (cellular compartment).

...
Additional arguments passed to specific methods.

verbose
Logical influencing whether messages (primarily missing GO terms arising during creation of the slim hierarchy) are reported.

Methods

idSrc="GOCollection", slimCollection="GOCollection", ontology="character", ..., verbose=FALSE

Classify idSrc GO terms into slimCollection categories. The hierarchy of terms included for each term is from the ontology (MF, BP, or CC) specified by ontology. verbose informs about, e.g., GO terms that are not found.

idSrc="ExpressionSet", slimCollection="GOCollection", ontology="character", ..., verbose=FALSE

Determine the (unique) GO terms implied by feature names in idSrc (using the annotation map identified in annotation(idSrc)).

Examples



myIds <- c("GO:0016564", "GO:0003677", "GO:0004345", "GO:0008265",
           "GO:0003841", "GO:0030151", "GO:0006355", "GO:0009664",
           "GO:0006412", "GO:0015979", "GO:0006457", "GO:0005618",
           "GO:0005622", "GO:0005840", "GO:0015935", "GO:0000311")
myCollection <- GOCollection(myIds)
fl <- system.file("extdata", "goslim_plant.obo", package="GSEABase")
slim <- getOBOCollection(fl)
goSlim(myCollection, slim, "MF")
data(sample.ExpressionSet)
goSlim(sample.ExpressionSet, slim, "MF", evidenceCode="TAS")

[Package GSEABase version 1.44.0 Index]