mapIdentifiers-methods {GSEABase}	R Documentation

Methods for Function mapIdentifiers in Package ‘GSEABase’

Description

These methods convert the genes identifiers of a gene set from one type to another, e.g., from EntrezIdentifier to AnnotationIdentifier. Methods can be called directly by the user; geneIdType<- provides similar functionality. verbose=TRUE produces warning messages when maps between identifier types are not 1:1, or a map has to be constructed on the fly (this situation does not apply when using the DBI-based annotation packages).

Methods

The following methods are defined on what="GeneSet":

what = "ANY", to = "ANY", from = "ANY", verbose=FALSE

This method warns of attempts to map from and to the same type, or generates an error if no suitable mapIdentifiers methods are available.

what = "GeneSet", to = "GeneIdentifierType", from = "missing", verbose=FALSE

This method will re-dispatch to a method with signature signature(what=what, to=to, from=geneIdType(what)), and is present so that a user can call mapIdentifiers without providing an explicit from argument.

what = "GeneSet", to = "GeneIdentifierType", from = "NullIdentifier", verbose=FALSE

This maps a gene set from gene identifiers represented by the NullIdentifier type (i.e., no type associated with the genes) to gene identifiers represent by any class derived from GeneIdentifierType.

what = "GeneSet", to = "NullIdentifier", from = "GeneIdentifierType", verbose=FALSE

This maps a gene set from gene identifiers represented by any GeneIdentifierType type to one represented by the NullIdentifier (i.e., no type associated with the genes).

what = "GeneSet", to = "GeneIdentifierType", from = "environment", verbose=FALSE

Maps identifiers found in what to the type described by to, using the map (key-value pairs) found in from.

what = "GeneSet", to = "GeneIdentifierType", from = "AnnDbBimap", verbose=FALSE

Maps identifiers found in what to the type described by to, using the map (key-value pairs) found in from.

The following methods are defined for what=GeneColorSet. These methods map gene- and phenotype color appropriately, and fail if coloring of gene identifiers involved in several-to-1 mappings conflict.

what = "GeneColorSet", to = "GeneIdentifierType", from = "missing", verbose=FALSE

This method will re-dispatch to a method with signature signature(what=what, to=to, from=geneIdType(what)), and is present so that a user can call mapIdentifiers without providing an explicit from argument.

what = "GeneColorSet", to = "GeneIdentifierType", from = "NullIdentifier", verbose=FALSE

This maps a gene set from gene identifiers represented by the NullIdentifier type (i.e., no type associated with the genes) to gene identifiers represent by any class derived from GeneIdentifierType.

what = "GeneColorSet", to = "NullIdentifier", from = "GeneIdentifierType", verbose=FALSE

This maps a gene set from gene identifiers represented by any GeneIdentifierType type to one represented by the NullIdentifier (i.e., no type associated with the genes).

what = "GeneColorSet", to = "GeneIdentifierType", from = "environment", verbose=FALSE

This method is not implemented, and exists to stop incorrect application of the GeneSet method.

what = "GeneColorSet", to = "GeneIdentifierType", from = "AnnDbBimap", verbose=FALSE

This method is not implemented, and exists to stop incorrect application of the GeneSet method.

A method exists for what="GeneSetCollection":

what = "GeneSetCollection", to = "GeneIdentifierType", from = "missing", verbose = FALSE

Map each gene set in what to gene identifier type to, using methods described above.

[Package GSEABase version 1.44.0 Index]